Remove 'ad' subpackage and move autodiff module to main boomdiff package; update test suite to reflect changes to package structure
