--- boomdiff/ad/__init__.py
@@ -1,1 +1,0 @@
-from .autodiff import AD
