+++ tests/__init__.py
@@ -1,3 +1,2 @@
 import boomdiff
-import boomdiff.ad
 import boomdiff.optimize
